Adding a dependency on XML Builder ('xmlbuilder').
Updating Surefire formatter to use this rather than trying to inline
building of XML structure.

Reverting formattedLog that was inserted (not needed).

Changed the naming convention on boolean flags to be more readable & to
make sure that properties aren't being set to false (which would be a
default value if not set).

Make sure it lints!